Don't assume that id_token response contains the authorization code
I was getting the following error after receiving the ID token from an
instance of the Ipsilon identity server.

    AssertionError: 'undefined' == 'string',
    {"name":"AssertionError","actual":"undefined","expected":"string","operator":"==","message":"'undefined'
    == 'string'","generatedMessage":true}

The issue is in verifying the 'c_hash' field of the response. The code
tried to check it against the 'code' field of the id_token, but this
isn't included in the response from Ipsilon.

I can't find anything in the OpenID Connect Core standard that requires
the authorization code to be included in the ID token response, so it
looks like the client library needs to be fixed to not require that.